Move split() and chunk() out of stdlib::nn
